change bundle name and location, moved to common
