+++ common/plugins/eu.esdihumboldt.hale.common.service.helper.test/.settings/edu.umd.cs.findbugs.core.prefs
