+++ common/plugins/eu.esdihumboldt.hale.common.service.helper.test/.settings/org.eclipse.core.resources.prefs
