+++ common/plugins/eu.esdihumboldt.hale.common.service.helper.test/.settings/org.eclipse.jdt.core.prefs
