+++ common/plugins/eu.esdihumboldt.hale.common.service.helper.test/build.properties
