+++ common/plugins/eu.esdihumboldt.hale.common.service.helper/.classpath
