+++ common/plugins/eu.esdihumboldt.hale.common.service.helper/.settings/edu.umd.cs.findbugs.core.prefs
