+++ common/plugins/eu.esdihumboldt.hale.common.service.helper/.settings/org.eclipse.jdt.core.prefs
