+++ common/plugins/eu.esdihumboldt.hale.common.service.helper/build.properties
