--- ui/plugins/eu.esdihumboldt.hale.ui.service.helper.test/META-INF/MANIFEST.MF
@@ -1,16 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: HALE UI service helper test
-Bundle-SymbolicName: eu.esdihumboldt.hale.ui.service.helper.test
-Bundle-Version: 3.1.0.qualifier
-Bundle-Vendor: wetransform GmbH
-Fragment-Host: eu.esdihumboldt.hale.ui.service.helper;bundle-version="3.1.0"
-Bundle-RequiredExecutionEnvironment: JavaSE-1.7
-Require-Bundle: groovy;bundle-version="2.3.7",
- org.junit;bundle-version="4.12.0"
-Import-Package: eu.esdihumboldt.hale.common.core.io.supplier,
- eu.esdihumboldt.hale.common.filter,
- eu.esdihumboldt.hale.common.instance.groovy,
- eu.esdihumboldt.hale.common.schema,
- eu.esdihumboldt.hale.common.schema.groovy,
- eu.esdihumboldt.util.groovy.paths
