--- ui/plugins/eu.esdihumboldt.hale.ui.service.helper/META-INF/MANIFEST.MF
@@ -1,12 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: Hale UI Service Implementer
-Bundle-SymbolicName: eu.esdihumboldt.hale.ui.service.helper
-Bundle-Version: 3.1.0.qualifier
-Bundle-RequiredExecutionEnvironment: JavaSE-1.7
-Bundle-Vendor: wetransform GmbH
-Import-Package: eu.esdihumboldt.hale.common.align.model,
- eu.esdihumboldt.hale.common.align.model.impl,
- eu.esdihumboldt.hale.common.instance.model,
- eu.esdihumboldt.hale.common.schema.model
-Export-Package: eu.esdihumboldt.hale.ui.service.helper.population
