+++ ui/plugins/eu.esdihumboldt.hale.ui/META-INF/MANIFEST.MF
@@ -83,6 +83,7 @@ Import-Package: com.google.common.base;version="1.6.0",
  eu.esdihumboldt.hale.common.schema.model.constraint.type,
  eu.esdihumboldt.hale.common.schema.model.impl,
  eu.esdihumboldt.hale.common.scripting.scripts.groovy,
+ eu.esdihumboldt.hale.common.service.helper.population,
  eu.esdihumboldt.hale.io.xsd.constraint,
  eu.esdihumboldt.hale.ui.common,
  eu.esdihumboldt.hale.ui.common.definition,
@@ -100,7 +101,6 @@ Import-Package: com.google.common.base;version="1.6.0",
  eu.esdihumboldt.hale.ui.filter.extension,
  eu.esdihumboldt.hale.ui.geometry.service,
  eu.esdihumboldt.hale.ui.geometry.service.impl,
- eu.esdihumboldt.hale.ui.service.helper.population,
  eu.esdihumboldt.hale.ui.util,
  eu.esdihumboldt.hale.ui.util.dialog,
  eu.esdihumboldt.hale.ui.util.handler,
